Fix getSymbol() to work with scopes
It was iterating through the list backwards, when it
should have been forwards (since we add inner scopes to
the front of the list).

Since we have working scoping, we no longer avoid name
collisions in templates. This commit updates templates to
shadow the name "test" where it makes sense.